remove ouble from cast and set as float in order to avoi php 8.5 deprecated warning

+++ src/ModelCollection.php
@@ -39,7 +39,7 @@ class ModelCollection extends Collection
                 if (is_bool($var)) {
                     $var = (bool) $var;
                 } else if (is_double($var)) {
-                    $var = (double) $var;
+                    $var = (float) $var;
                 } else if (is_float($var)) {
                     $var = (float) $var;
                 } else if (is_bool($var) || is_numeric($var)) {
